langchain-python-webapi/requirements.txt
fastapi==0.111.0
langchain==0.1.17
azure-identity==1.16.0
langchain-openai==0.1.6
python-dotenv==1.0.1
langchainhub==0.1.15
https://gist.github.com/anthonychu/f860806c7676c41a737ac8eda893ad25/raw/c1e6b21c40bff8a4117151f897f3539790d8216d/langchain_azure_dynamic_sessions-0.0.1-py3-none-any.whl